app restructure
