+++ packages/app/src/App.tsx
@@ -9,7 +9,7 @@ import { CssBaseline } from "@mui/material";
 import { ThemeProvider } from "@mui/material/styles";
 import * as React from "react";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
-import { HeaderProps } from "./components/Header";
+import { HeaderProps } from "./components/layout/Header/Header";
 import theme from "./theme";
 
 /**
@@ -17,13 +17,13 @@ import theme from "./theme";
  * React.lazy and dynamic import() statements for lazy loading React
  * components.
  */
-const HomePage = React.lazy(() => import("./components/HomePage"));
-const MainPage = React.lazy(() => import("./components/MainPage"));
-const ModelPage = React.lazy(() => import("./components/ModelPage"));
-const PackagePage = React.lazy(() => import("./components/PackagePage"));
-const ProjectPage = React.lazy(() => import("./components/ProjectPage"));
+const HomePage = React.lazy(() => import("./components/pages/HomePage/HomePage"));
+const MainPage = React.lazy(() => import("./components/layout/MainPage/MainPage"));
+const ModelPage = React.lazy(() => import("./components/pages/ModelPage/ModelPage"));
+const PackagePage = React.lazy(() => import("./components/pages/PackagePage/PackagePage"));
+const ProjectPage = React.lazy(() => import("./components/pages/ProjectPage/ProjectPage"));
 const RouteError = React.lazy(() => import("./components/RouteError"));
-const WorkbookPage = React.lazy(() => import("./components/WorkbookPage"));
+const WorkbookPage = React.lazy(() => import("./components/pages/WorkbookPage/WorkbookPage"));
 
 // Create router configuration function
 export const createMalloyRouter = (
