+++ packages/app/src/components/layout/Header/Header.tsx
@@ -14,7 +14,7 @@ import {
 } from "@mui/material";
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import BreadcrumbNav from "./BreadcrumbNav";
+import BreadcrumbNav from "../BreadcrumNav/BreadcrumbNav";
 
 export interface HeaderProps {
    logoHeader?: React.ReactElement;
