+++ packages/app/src/components/layout/index.ts
@@ -1,0 +1,4 @@
+export * from "./Header";
+export type { HeaderProps } from "./Header";
+export { default as MainPage } from "./MainPage/MainPage";
+export { default as BreadcrumbNav } from "./BreadcrumNav/BreadcrumbNav";
