+++ packages/app/src/components/pages/index.ts
@@ -1,0 +1,5 @@
+export { default as HomePage } from "./HomePage/HomePage";
+export { default as ProjectPage } from "./ProjectPage/ProjectPage";
+export { default as PackagePage } from "./PackagePage/PackagePage";
+export { default as ModelPage } from "./ModelPage/ModelPage";
+export { default as WorkbookPage } from "./WorkbookPage/WorkbookPage";
