+++ packages/app/src/index.ts
@@ -1,10 +1,10 @@
 // Main App exports
-export { MalloyPublisherApp, createMalloyRouter } from "./App";
+export { createMalloyRouter, MalloyPublisherApp } from "./App";
 export type { MalloyPublisherAppProps } from "./App";
 
 // Theme exports
+export * from "./components/layout/Header";
 export { default as theme } from "./theme";
-export * from "./components";
 // Additional component exports for advanced usage
-export { default as BreadcrumbNav } from "./components/BreadcrumbNav";
-export { default as ToggleColorMode } from "./components/ToggleColorMode";
+export { default as BreadcrumbNav } from "./components/layout/BreadcrumNav/BreadcrumbNav";
+export { default as ToggleColorMode } from "./components/theme/ToggleColorMode/ToggleColorMode";
